Add deadline and notification logic to app
The commit refactors the main app file to include new logic for handling deadlines and sending notifications. The "App" struct was introduced containing the actual database. "Run" and "onReady" were also updated to be methods of this struct. Various libraries have been imported to support these changes. Other updates in this commit include addition of notification functionality when the "Notify" system tray menu item is clicked and addition of a new "SetDeadline" menu item that displays a GUI for setting a deadline when clicked.