Remove unused utils package and unnecessary Github button from about page
The utils package was imported in about.go, but not used in the actual code, so it was removed to clean up unnecessary imports. The Github button that used to open our Github page when clicked has been removed from the zenity.Info window; this decision was taken to simplify the About page, redirecting users to our Github page was deemed unnecessary. Zenity.InfoIcon now stands alone for a cleaner look.